requirements.txt
flask==3.0.2
numpy==1.26.4
pillow==10.2.0
